+++ app/src/main/java/cy/volleybolley/courts/presentation/model/CourtUiMapper.kt
@@ -3,8 +3,13 @@ package cy.volleybolley.courts.presentation.model
 import android.location.Location
 import com.google.android.gms.maps.model.LatLng
 import cy.volleybolley.courts.domain.model.Court
+import java.util.Locale
 
 object CourtUiMapper {
+    private const val METERS_IN_KILOMETER = 1000f
+    private const val DISTANCE_FORMAT = "%.1f km"
+    private const val DISTANCE_UNKNOWN = "-- km"
+
     fun Court.toCourtUI(userLocation: LatLng? = null): CourtUi {
         val distanceText = userLocation?.let { userLoc ->
             val results = FloatArray(1)
@@ -13,8 +18,8 @@ object CourtUiMapper {
                 this.location.latitude, this.location.longitude,
                 results
             )
-            "${(results[0] / 1000).toInt()} km"
-        } ?: "-- km"
+            String.format(Locale.US, DISTANCE_FORMAT, results[0] / METERS_IN_KILOMETER)
+        } ?: DISTANCE_UNKNOWN
 
         return CourtUi(
             courtId = this.courtId,
@@ -36,8 +41,8 @@ object CourtUiMapper {
                 this.location.latitude, this.location.longitude,
                 results
             )
-            "${(results[0] / 1000).toInt()} km"
-        } ?: "-- km"
+            String.format(Locale.US, DISTANCE_FORMAT, results[0] / METERS_IN_KILOMETER)
+        } ?: DISTANCE_UNKNOWN
 
         return this.copy(distanceText = newDistance)
     }
