+++ app/src/main/res/values/strings.xml
@@ -258,4 +258,5 @@
     <string name="default_name">No name</string>
     <string name="default_level">LIGHT</string>
     <string name="navigation">Navigation</string>
+    <string name="start_position">You are here</string>
 </resources>
